requirements.txt
torch>=2.2.0
torchvision>=0.17.0
nuscenes-devkit
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.5.2
transformers>=4.30.0
timm>=0.6.13
opencv-python>=4.7.0
scikit-learn>=1.2.0
tqdm>=4.65.0
wandb>=0.15.0
seaborn>=0.11.0
PyYAML>=5.4.0